Move dictutil tests out.
